Extract time helpers from engine_controller_misc to efitime.cpp
