+++ firmware/controllers/engine_controller.cpp
@@ -163,21 +163,6 @@ class EngineStateBlinkingTask : public PeriodicTimerController {
 
 static EngineStateBlinkingTask engineStateBlinkingTask;
 
-/**
- * 32 bit return type overflows in 23(or46?) days. tag#4554. I think we do not expect rusEFI to run for 23 days straight days any time soon?
- */
-efitimems_t currentTimeMillis(void) {
-	return US2MS(getTimeNowUs());
-}
-
-/**
- * Integer number of seconds since ECU boot.
- * 31,710 years - would not overflow during our life span.
- */
-efitimesec_t getTimeNowSeconds(void) {
-	return getTimeNowUs() / US_PER_SECOND;
-}
-
 static void resetAccel() {
 	engine->tpsAccelEnrichment.resetAE();
