+++ firmware/util/util.mk
@@ -5,6 +5,7 @@ UTILSRC = \
 
 UTILSRC_CPP = \
 	$(UTIL_DIR)/histogram.cpp \
+	$(UTIL_DIR)/efitime.cpp \
 	$(UTIL_DIR)/containers/listener_array.cpp \
 	$(UTIL_DIR)/containers/local_version_holder.cpp \
 	$(UTIL_DIR)/math/biquad.cpp \
